test(server): strengthen budget RBAC coverage
